Refactor `Format` into separate mod
